+++ bloom_nofos/nofos/templatetags/get_icon_for_section.py
@@ -2,9 +2,9 @@ from django import template
 
 register = template.Library()
 
-from ..nofo import get_icon_for_section
+from .utils import get_icon_for_section as get_icon_for_section_func
 
 
 @register.filter()
-def get_icon_for_section(section_name="", theme="blue"):
-    return get_icon_for_section(section_name, theme)
+def get_icon_for_section(section_name=""):
+    return get_icon_for_section_func(section_name)